refactor common logic
