+++ apps/modernization-ui/src/apps/patient/add/cancelAddPatientPanel/index.ts
@@ -1,2 +1,3 @@
 export { useShowCancelModal } from './useShowCancelModal';
 export { CancelAddPatientPanel } from './CancelAddPatientPanel';
+export { handleNativeCancelAddPanel } from './handleNativeCancelAddPanel';
