+++ apps/modernization-ui/src/apps/patient/add/extended/AddPatientExtended.tsx
@@ -10,7 +10,7 @@ import { useBasicExtendedTransition } from 'apps/patient/add/useBasicExtendedTra
 import { AddPatientExtendedInPageNav } from './nav/AddPatientExtendedNav';
 import { ExtendedNewPatientEntry } from './entry';
 import { AddPatientExtendedForm } from './AddPatientExtendedForm';
-import { CancelAddPatientPanel, useShowCancelModal } from '../cancelAddPatientPanel';
+import { CancelAddPatientPanel, useShowCancelModal, handleNativeCancelAddPanel } from '../cancelAddPatientPanel';
 import { useAddPatientExtendedDefaults } from './useAddPatientExtendedDefaults';
 import { useAddExtendedPatient } from './useAddExtendedPatient';
 import { AddExtendedPatientInteractionProvider } from './useAddExtendedPatientInteraction';
@@ -70,16 +70,7 @@ export const AddPatientExtended = () => {
     }, [interaction.status]);
 
     useEffect(() => {
-        const handleBeforeUnload = (e: BeforeUnloadEvent) => {
-            if (form.formState.isDirty && !form.formState.isSubmitted) {
-                e.preventDefault();
-            }
-        };
-
-        window.addEventListener('beforeunload', handleBeforeUnload);
-        return () => {
-            window.removeEventListener('beforeunload', handleBeforeUnload);
-        };
+        handleNativeCancelAddPanel(form.formState.isDirty, form.formState.isSubmitted);
     }, [form.formState.isDirty, form.formState.isSubmitted]);
 
     return (
